second push

+++ new_pro/shorten/views.py
@@ -1,13 +1,17 @@
 from django.shortcuts import render, get_object_or_404
+# To return an HttpResponse
 from django.http import HttpResponse, HttpResponseRedirect
 from .models import bitly
 from .form import bitlyform, editBitly 
 from .util import create_shortcode
 import json
 from datetime import datetime
+from django.urls import reverse
+# to use it as the dynamic urls
+from django.contrib.auth.decorators import login_required
 
 def index(request):
-    objects= bitly.objects.all()
+    objects= bitly.objects.all()[: :-1]
     print(objects)
 
     context={'objs': objects}
@@ -22,7 +26,7 @@ def create(request):
         instance.datewise = "{}"
         instance.save()
 
-        return HttpResponseRedirect("http://127.0.0.1:8000/home")
+        return HttpResponseRedirect(reverse('index'))
 
     context={"urlform":form}
     return render(request,"create.html",context)
@@ -41,19 +45,22 @@ def goto(request,shortcode=None):
     return HttpResponseRedirect(qs.long_url)
 
 def update(request,pk=None):
-    qs=get_object_or_404(bitly,id=pk)
-    form=editBitly(request.POST or None,instance=qs)
+    if request.user.is_authenticated:
+        qs=get_object_or_404(bitly,id=pk)
+        form=editBitly(request.POST or None,instance=qs)
 
-    if form.is_valid():
-        form.save()
-        return HttpResponseRedirect("http://127.0.0.1:8000/home")
-    context={'urlform':form}
-    return render(request,"create.html",context)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('index'))
+        context={'urlform':form}
+        return render(request,"create.html",context)
+    return HttpResponseRedirect(reverse('index'))
 
 def delete(request,pk=None):
-    qs=get_object_or_404(bitly,id=pk)
-    qs.delete()
-    return HttpResponseRedirect("http://127.0.0.1:8000/home")
+    if request.user.is_authenticated:
+        qs=get_object_or_404(bitly,id=pk)
+        qs.delete()
+    return HttpResponseRedirect(reverse('index'))
 
 
 # Create your views here.
